src/sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ossa-sdk"
version = "0.5.0"
description = "Python SDK for the Open Standard for Software Agents (OSSA)"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "BlueFly.io", email = "agents@bluefly.io" },
]
keywords = ["ossa", "agents", "ai", "mcp", "a2a", "manifest", "governance", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.39.0"]
openai = ["openai>=1.0"]
google = ["google-generativeai>=0.8.0"]
langchain = ["langchain-core>=0.3.0"]
crewai = ["crewai>=0.80.0"]
all-providers = [
    "anthropic>=0.39.0",
    "openai>=1.0",
    "google-generativeai>=0.8.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "mypy>=1.8",
    "ruff>=0.4",
]

[project.scripts]
ossa-py = "ossa.cli:main"

[project.urls]
Homepage = "https://openstandardagents.org"
Documentation = "https://openstandardagents.org/docs/sdks/python-sdk"
Repository = "https://gitlab.com/blueflyio/ossa/openstandardagents"

[tool.hatch.build.targets.wheel]
packages = ["ossa"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
